moved types to types folder
