+++ src/app/GlobalRedux/Features/Data/dataSlice.ts
@@ -1,6 +1,6 @@
 'use client';
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
-import { MarketData } from "../../../../../MarketData";
+import { MarketData } from "../../../types/MarketData";
 
 
 export const fetchData = createAsyncThunk(
