+++ src/app/types/MarketData.ts
