+++ src/app/types/globalTypes.ts
